LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-createAnswer-expected.txt


PASS createAnswer() with null remoteDescription should reject with InvalidStateError 
FAIL createAnswer() after setting remote description should succeed assert_equals: Expect m=video line to be present in generated SDP expected 1 but got 0
PASS createAnswer() when connection is closed reject with InvalidStateError 
PASS createAnswer() when connection is closed in parallel should never resolve 
